refactor: move Mermaid components into dedicated directory module
