+++ client/src/components/Messages/Content/Mermaid/MermaidErrorBoundary.tsx
@@ -2,7 +2,6 @@ import React from 'react';
 
 interface MermaidErrorBoundaryProps {
   children: React.ReactNode;
-  /** The mermaid code to display as fallback */
   code: string;
 }
 
@@ -10,10 +9,6 @@ interface MermaidErrorBoundaryState {
   hasError: boolean;
 }
 
-/**
- * Error boundary specifically for Mermaid diagrams.
- * Falls back to displaying the raw mermaid code if rendering fails.
- */
 class MermaidErrorBoundary extends React.Component<
   MermaidErrorBoundaryProps,
   MermaidErrorBoundaryState
@@ -32,7 +27,6 @@ class MermaidErrorBoundary extends React.Component<
   }
 
   componentDidUpdate(prevProps: MermaidErrorBoundaryProps) {
-    // Reset error state if code changes (e.g., user edits the message)
     if (prevProps.code !== this.props.code && this.state.hasError) {
       this.setState({ hasError: false });
     }
@@ -42,10 +36,10 @@ class MermaidErrorBoundary extends React.Component<
     if (this.state.hasError) {
       return (
         <div className="w-full overflow-hidden rounded-md border border-border-light">
-          <div className="rounded-t-md bg-gray-700 px-4 py-2 font-sans text-xs text-gray-200">
+          <div className="rounded-t-md bg-surface-secondary px-4 py-2 font-sans text-xs text-text-secondary">
             {'mermaid'}
           </div>
-          <pre className="overflow-auto whitespace-pre-wrap rounded-b-md bg-gray-900 p-4 font-mono text-xs text-gray-300">
+          <pre className="overflow-auto whitespace-pre-wrap rounded-b-md bg-surface-primary-alt p-4 font-mono text-xs text-text-secondary">
             {this.props.code}
           </pre>
         </div>
