+++ client/src/components/Messages/Content/Mermaid/index.ts
@@ -1,0 +1,2 @@
+export { default } from './Mermaid';
+export { default as MermaidErrorBoundary } from './MermaidErrorBoundary';
